+++ hiveeyes/datalogger.py
@@ -43,19 +43,19 @@ class HiveeyesDatalogger(TerkinDatalogger):
         try:
             self.add_hx711_sensor()
         except Exception as ex:
-            log.exception('Skipping HX711 sensor')
+            log.exception('Setting up HX711 sensor failed')
 
         # Setup the DS18X20.
         try:
             self.add_ds18x20_sensor()
         except Exception as ex:
-            log.exception('Skipping DS18x20 sensor')
+            log.exception('Setting up DS18X20 sensor failed')
 
         # Setup the BME280.
         try:
             self.add_bme280_sensor()
         except Exception as ex:
-            log.exception('Skipping bme280 sensor')
+            log.exception('Setting up BME280 sensor failed')
 
     def add_hx711_sensor(self):
         """
@@ -64,6 +64,10 @@ class HiveeyesDatalogger(TerkinDatalogger):
 
         # Initialize HX711 sensor component.
         settings = self.settings.get('sensors.registry.hx711')
+
+        if settings.get('enabled') is False:
+            log.info("Skipping HX711 device on pins {}/{}".format(settings['pin_dout'], settings['pin_pdsck']))
+            return
 
         hx711_sensor = HX711Sensor()
         hx711_sensor.set_address(settings.get('address', 0x00))
@@ -90,7 +94,7 @@ class HiveeyesDatalogger(TerkinDatalogger):
         settings = self.settings.get('sensors.registry.ds18x20')
 
         bus = self.sensor_manager.get_bus_by_name(settings['bus'])
-        sensor = DS18X20Sensor()
+        sensor = DS18X20Sensor(settings=settings)
         sensor.acquire_bus(bus)
 
         # Start sensor.
@@ -106,6 +110,10 @@ class HiveeyesDatalogger(TerkinDatalogger):
 
         settings = self.settings.get('sensors.registry.bme280')
         bus = self.sensor_manager.get_bus_by_name(settings['bus'])
+
+        if settings.get('enabled') is False:
+            log.info("Skipping BME280 device {} on bus {}".format(hex(settings['address']), bus.name))
+            return
 
         sensor = BME280Sensor()
         if 'address' in settings: